Info fix: Borrado de partidos antes de insertarlos

+++ app/src/main/java/gm/mobi/android/db/manager/WatchManager.java
@@ -80,8 +80,8 @@ public class WatchManager extends AbstractManager{
         }
     }
 
-    public void deleteAllWatches() {
-        db.execSQL("DELETE FROM " + GMContract.WatchTable.TABLE);
+    public void deleteAllWatches(Long exceptUserId) {
+        db.execSQL("DELETE FROM " + GMContract.WatchTable.TABLE+ " WHERE "+WatchTable.ID_USER+" = "+exceptUserId);
     }
 
     public void insertInSync(){
